EPBR-1480: Adding in the missing view model methods to NI data export
These nodes were originally created as part of the Open Data Communities export, so just needed to copy and paste them over to the report_from_ni_assessment_xml. However, most of the work taken in this commit was in the test expectations themselves, as we hadn't previously tested the NI schemas for ODE. This required going through all of the failing tests, checking the schemas, view models and fixtures.