Mechanical-Ventilation update the enum method for pre 12 schemas
We were not returning the correct xml values for rdsap pre 12 schemas for mechanical ventilation. 
This was not being tested for so added a unit test which passed but then the sap wrapper integration test would fail when I updated the different_fields object to reflect the different value that should have been returned but currently wasn't.

This highlighted that the conditional check in the mechanical ventilation method in the enums was not reading the schema version correctly as we were passing a symbol to it yet the condition was checking for strings.