Estimated-Energy-Cost is not an element in any XML schema
Remove the attempted xpath lookup for an <Estimated-Energy-Cost> element from all view models.  This does not exist in any schema - estimated_energy_cost is calculated as the sum of lighting_cost_current, heating_cost_current and hot_water_cost_current.

I've left the declaration of the field in the view models, as there are other calculated fields that have their existence declared like this (e.g. potential_energy_saving).  It's not clear to me whether this is necessary or useful, but they aren't doing any harm there and things are less likely to break in unexpected ways if we leave them in there, at least for now.