Read Asset-Rating and other ratings as integers
These fields are defined as integers in the XML schema
and we should be processing them as numbers rather than
strings.  (We currently have a mix - domestic ratings are
integers but non-domestic are strings).

Have added the &. operator to calls to to_i, as nil.to_i
returns zero, which is not what we want (nil should be nil,
and an explicit zero rating is zero).