Don't use explicit nil returns for RdSAP 10.2 schema elements
Rather than having to define explicit nil returns in the view
model for every schema that doesn't happen to contain a
particular element we can instead check at the time of use
whether the view model for the schema "responds to" the method
to fetch the field.

Taking this approach will make it much easier to add new fields
in future new schemas, as we won't need to go through all
historical schemas to add an explicit nil return.